1.13 update & verify command
