+++ gradle.properties
@@ -1,0 +1,2 @@
+pluginGroup=com.frostcraft
+pluginVersion=1.0
