--- src/com/frostcraft/admintools/Main.java
@@ -1,28 +1,0 @@
-package com.frostcraft.admintools;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.UUID;
-
-import org.bukkit.ChatColor;
-import org.bukkit.plugin.java.JavaPlugin;
-
-public class Main extends JavaPlugin {
-	public static boolean enablePhysics = true;
-	public static List<UUID> vanishList = new ArrayList<UUID>(); //List of vanished players
-	public static List<UUID> muteList = new ArrayList<UUID>(); //List of muted players
-
-	//Plugin enable
-	public void onEnable() {
-		getServer().getPluginManager().registerEvents(new BlockListener(),this);
-		getServer().getPluginManager().registerEvents(new PlayerListener(),this);
-		
-		getCommand("fcadmin").setExecutor(new FCAdminCommand(this));
-		
-		this.getServer().getConsoleSender().sendMessage(ChatColor.AQUA + "[FrostCraft-AdminTools] Plugin enabled.");
-	}
-	
-	//Plugin disable
-	public void onDisable() {
-		this.getServer().getConsoleSender().sendMessage(ChatColor.AQUA + "[FrostCraft-AdminTools] Plugin disabled.");
-	}
-}
